Chatfix

+++ trade/views.py
@@ -108,38 +108,52 @@ def create(request):
 @login_required
 def update(request, pk):
     trade = Trades.objects.get(pk=pk)
-    photos = trade.photo_set.all()
-    instancetitle = trade.title
-    if request.method == "POST":
-        review_form = CreateTrade(request.POST, request.FILES, instance=trade)
-        if photos:
-            photo_form = PhotoForm(request.POST, request.FILES, instance=photos[0])
-        else:
-            photo_form = PhotoForm(request.POST, request.FILES)
-        images = request.FILES.getlist("image")
-        for photo in photos:
-            if photo.image:
-                photo.delete()
-        kb = Keyboard.objects.get(name=request.POST["keyboard"])
-        if review_form.is_valid() and photo_form.is_valid():
-            trade = review_form.save(commit=False)
-            trade.user = request.user
-            trade.keyboard = kb
-            if len(images):
-                for image in images:
-                    image_instance = Photo(trade=trade, image=image)
-                    trade.save()
-                    image_instance.save()
+    if request.user == trade.user:
+        photos = trade.photo_set.all()
+        instancetitle = trade.title
+        if request.method == "POST":
+            review_form = CreateTrade(request.POST, request.FILES, instance=trade)
+            if photos:
+                photo_form = PhotoForm(request.POST, request.FILES, instance=photos[0])
             else:
-                trade.save()
-            return redirect("trade:index")
-    else:
-        review_form = CreateTrade(instance=trade)
-        if photos:
-            photo_form = PhotoForm(instance=photos[0])
+                photo_form = PhotoForm(request.POST, request.FILES)
+            images = request.FILES.getlist("image")
+            for photo in photos:
+                if photo.image:
+                    photo.delete()
+            kb = Keyboard.objects.get(name=request.POST["keyboard"])
+            if review_form.is_valid() and photo_form.is_valid():
+                trade = review_form.save(commit=False)
+                trade.user = request.user
+                trade.keyboard = kb
+                if len(images):
+                    for image in images:
+                        image_instance = Photo(trade=trade, image=image)
+                        trade.save()
+                        image_instance.save()
+                else:
+                    trade.save()
+                return redirect("trade:index")
+        else:
+            review_form = CreateTrade(instance=trade)
+            if photos:
+                photo_form = PhotoForm(instance=photos[0])
+            else:
+                photo_form = PhotoForm()
+        if request.user.is_authenticated:
+            new_message = Notification.objects.filter(
+                Q(user=request.user) & Q(check=False)
+            )
+            message_count = len(new_message)
+            context = {
+                "count": message_count,
+                "review_form": review_form,
+                "photo_form": photo_form,
+                "instancetitle": instancetitle,
+                "trade": trade,
+            }
         else:
             photo_form = PhotoForm()
-
     if request.user.is_authenticated:
         new_message = Notification.objects.filter(Q(user=request.user) & Q(check=False))
         message_count = len(new_message)
